requirements.txt
# Keep this file in alphabetical order.  No logical grouping of blocks.

Django==1.11.5
celery==4.0.2
django-compressor==2.1.1
django-jsonfield==1.0.1
django-model-utils==2.6.1
django-tables2==1.2.9
django-mysql==2.1.0
git+ssh://git@jp74.git.beanstalkapp.com/django-ecom6.git@1.6#egg=django-ecom6==1.6
googleads==7.0.0
mysqlclient==1.3.9
python-dateutil==2.6.1
redis==2.10.5
requests==2.18.4
simplejson==3.11.1
